Move quote cost into shared location
